requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
datasets==4.5.0
dill==0.4.0
filelock==3.21.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
kornia==0.8.2
kornia_rs==0.1.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
networkx==3.6.1
numpy==2.4.2
onedrivedownloader==1.1.3
pandas==3.0.0
pillow==12.1.1
propcache==0.4.1
pyarrow==23.0.0
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.2
safetensors==0.7.0
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
timm==0.9.8
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
typer==0.23.0
typer-slim==0.23.0
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
wheel==0.46.3
xxhash==3.6.0
yarl==1.22.0
